feat: introduce statement blocks and variables with semantic analysis
Statement blocks allow to execute multiple statements in a row. On top of that,
variables can used in a program. Lax supports variable declaration, assignment,
and the use of variables in expressions.

A new layer is added after parsing and before interpreting for semantic
analysis. During this phase, type checking is performed to ensure that
computations and values assigned to variables are type-consistant. On top of
that, the semantic analyzer checks if a variable is declared before being used
and that a variable is not declared twice in the same scope.