fix: create a new scope for conditional branches with a single statement
Single statements executed after a condition in a then/else branches kept the
the same scope. This resulted for example in a variable savec in the global
scope when it was declared in a else branch. This commit fixes this behavior,
and now a new scope is created whenever a then/else branch is entered.