feat: introduce functions
Add support for functions, allowing to define and call them.
Doing so executes the instructions defined in the function body. It is possible
to pass arguments to functions, as specified in the function signature, as well
as return values from function calls.

For this purpose, a new implementation for values in Lax has been designed.
Moreover, assignments are no longer considered as statements but as
expressions.

Functions semantics are not fully supported yet, so errors may occur when
manipulating them in source code.