!feat: implement VM and compiler to run bytecode from source code
BREAKING CHANGE
Change the way the source code is executed. Before that commit, the AST was
traversed at runtime to execute each node as code. Now, a compiler is
responsible of converting the AST to bytecode that can be run by a virtual
machine.

The print keyword is now abandonned, as well as the StmtPrintNode, and is
replaced by the native function print().

Another notable change is the addition of closures over functions. Variables
used in functions that go out of scope get their state saved in closures, so
that they can still be accessed when the function is called.