style: refactor code for improved readability
Code and comments have been adjusted to improve readability and keep the code
style consistent between files. Lines don't exceed the 80 character limit, the
ampersand and the asterisk for pointers and references are now stick to the
type, thus emphasizing the type of the pointer/reference.

Code indents have been transformed to 4 spaces instead of a tab character,
to allow better alignment for multi-line instructions.