docs: refactor documentation using doxygen
The comments and some code in various files have been cleaned up, refactored,
and edited with the intention of improving readability and better conveying
the meaning of the classes and functions.

This will help in understanding the flow and purpose of the code and also
enable developers to navigate and update code more efficiently.

The Doxygen documentation format has been chosen as it is natively supported
by JetBrains products, and because it provides a consistent format for
documentation.